Bib2HTML

Description

This module checks a BibTex database for correctness and outputs warning messages when it encounters inconsistensies. A syntactically valid, but semantically incorrect BibTex file does not halt the program, but does provide highly specific messages. After checking, the BibTex database is converted into an HTML ATerm. The module's main function uses a Component wrapped in ioWrap for the pipe-passing mode, and provides the ioCheckBibTex function for checking in file mode.

Documentation

main :: IO ()

This calls ioWrap pipeline.

Validation Code

Required and optional attribute generation

type ValidationListGen = EntryType -> [AttrType]

Type synonym for a generator of a list of AttrTypes for validation.

reqFields :: ValidationListGen

Given an EntryType, produce a list of required fields for that EntryType. Based upon http://nwalsh.com/tex/texhelp/bibtx-7.html.

NOTE: For the entry types, we did not take into account the following:

The Book Entry type requires either the Editor or the Author attribute type. We only put Author as a required field, and Editor as an optional field.

The Inbook Entry type does the same, as well as allowing the Chapter and / or Pages attribute type. We put Chapter as a required field, and Pages as an optional field.

optFields :: ValidationListGen

Given an EntryType, produce a list of optional fields for that EntryType. Based upon http://nwalsh.com/tex/texhelp/bibtx-7.html.

type WarningGenerator = String -> EntryType -> AttrType -> String

Type synonym for a function that takes some information for generating a meaningful warning message.

Warning generation

reqNotPresent :: WarningGenerator

Generates warnings for missing required attributes.

optNotPresent :: WarningGenerator

Generates warnings for missing optional attributes.

unusedAttrs :: WarningGenerator

Generates warnings for irrelevant (not required and not optional) attributes.

Attribute type checking

check

Arguments

:: [AttrType]
list of attributes

-> AttrType
attribute that will be checked if it is in the list

-> (Bool, AttrType)
resulting tuple

Given a list of AttrTypes and another AttrType, check if it is an element of the first argument and return an accoridng tuple.

type CheckFunction = EntryType -> [AttrType] -> [(Bool, AttrType)]

Type synonym for a function that checks the validity of an EntryType and its AttrTypes.

checkFor :: ValidationListGen -> CheckFunction

Abstracts over checking.

checkForReq, checkForUnused, checkForOpt :: CheckFunction

These functions check for required fields, optional fields and unused fields, respectively.

Clean-up functions

removeDoubles :: Ord a => [a] -> [a]

Converts a list to a Set and back again to get rid of doubles in the list.

removeUnused :: [(Bool, AttrType)] -> [Attr] -> [Attr]

Given a list of tuples and a list of Attrs, remove all Attrs whose AttrType is coupled with False in the tuple.

lookupAttr :: AttrType -> [Attr] -> Attr

Given an AttrType and a list of Attrs, find the corresponding Attr value.

Printing functions

checkForErrors :: Monad m => String -> EntryType -> [AttrType] -> m ()

Unused. Checks for required attributes and halts the program if there are missing required attributes.

printIssues :: WarningGenerator -> String -> EntryType -> [(Bool, AttrType)] -> Feedback ()

Abstracts over the printing of Strings generated by a WarningGenerator.

printMessages :: CheckFunction -> WarningGenerator -> String -> EntryType -> [AttrType] -> Feedback ()

Abstracts over attribute checking and warning message printing.

printReqErrors, printUnusedWarnings, printOptWarnings :: String -> EntryType -> [AttrType] -> Feedback ()

These functions print warning messages for required fields, optional fields and unused fields, respectively.

Getting/Setting AttrTypes

extractAttrTypes :: [Attr] -> [AttrType]

Extracts the AttrTypes from a list of Attrs.

recombineAttrTypes :: [Attr] -> [AttrType] -> [Attr]

Recombines a list of AttrTypes with their corresponding Attrs. Note that we could have used a Lens datatype here.

Validation And Ordering

checkAttributes :: Entry -> Feedback Entry

Checks the attributes of an Entry for required, optional and unused attributes, and removes duplicate attributes.

checkErrors :: Entry -> Feedback ()

Unused. Checks for required attributes and halts the program if there are missing required attributes.

orderAttributes :: [Attr] -> [Attr]

Orders attributes according to their order of appearance in the definition of AttrType.

orderEntries :: [Entry] -> [Entry]

Orders entries according to their order of appearance in the definition of EntryType, currently alphabetically.

IO Monad Code

ioCheckBibTex :: StrBibTex -> IO ATerm

Use this to check a BibTex database in the IO Monad.

Component Code

convertToAterm :: Component String ATerm

Converts a String to an ATerm using parser.

convertToBibTex :: Component ATerm BibTex

Converts an ATerm to a BibTex database using toTree.

checkBibTex :: Component BibTex BibTex

Checks a BibTex database for validity.

sortBibTex :: Component BibTex BibTex

Sorts a Bibtex database's Entry and Attr elements.

toHTMLBibTex :: Component BibTex String

Converts a BibTex database to an HTML ATerm, which is then converted to a String using show.

pipeline :: Component String String

A Component pipeline that binds all the Components mentioned above.

HTML Generation

bibTex2HTML :: BibTex -> ATerm

Converts a BibTex database to an HTML ATerm.

createLink :: Entry -> ATerm

Generates an <a> tag from an Entry.

createName :: Entry -> ATerm

Generates a name from an Entry.

createTable :: [ATerm] -> ATerm

Generates a <table> tag for a list of ATerms.

createTableRow :: [ATerm] -> ATerm

Generates a <tr> tag for a list of ATerms.

createTableField :: [ATerm] -> ATerm

Generates a <td> tag for a list of ATerms.

createEntryText :: Entry -> [ATerm]

Formats an Entry.

find :: [Attr] -> AttrType -> [ATerm]

Search through the given list of Attrs to find a corresponding find' function, and call it.

find' :: AttrType -> AttrValue -> [ATerm]

Formats an AttrValue.

commaFind, defaultFind :: String -> [ATerm]

Common formats.