BibTex

Description

This module contains the datatype declarations. There are two prominent datatypes: StrBibTex and BibTex. StrBibTex is used during parsing, and is more liberal in the values of the entry and atrribute types. BibTex only allows certain values for entry and attribute types using the EntryType and AttrType datatypes. Using fromTree and toTree, we convert between these two datatypes.

The BibTex datatype, used in checking and HTML generation

data BibTex

A complete BibTex file.

Constructors

BibTex [Entry]

data Entry

An Entry. Consists of the type of the Entry, the title of the Entry, and a list of attributes.

Constructors

Entry EntryType EntryTitle [Attr]

data EntryType

The type of an Entry.

Constructors

Article

Book

Booklet

Conference

Inbook

Incollection

Inproceedings

Manual

Mastersthesis

Misc

Phdthesis

Proceedings

Techreport

Unpublished

OtherEntry String

type EntryTitle = String

data Attr

An attribute of an Entry. Consists of the type of the attribute, and the attribute value.

Constructors

Attr AttrType AttrValue

data AttrType

The type of an Attribute.

Constructors

Author

Editor

Title

Booktitle

Howpublished

Type

Year

Journal

Month

Note

Key

Pages

Chapter

Volume

Number

Edition

Series

Organization

Publisher

School

Institution

Address

OtherAttr String

type AttrValue = String

generateToTreeAttrType :: [AttrType] -> [TreeParser AttrType]

A function that maps over the list of AttrTypes and produces an appropriate parser for each.

generateToTreeEntryType :: [EntryType] -> [TreeParser EntryType]

A function that maps over the list of EntryTypes and produces an appropriate parser for each.

The StrBibTex datatype, used in parsing

data StrBibTex

Constructors

StrBibTex [StrEntry]

data StrEntry

Constructors

StrEntry StrEntryType StrEntryTitle [StrAttr]

data StrEntryType

Constructors

StrEntryType String

type StrEntryTitle = String

data StrAttr

Constructors

StrAttr StrAttrType StrAttrValue

data StrAttrType

Constructors

StrAttrType String

type StrAttrValue = String

uflr :: String -> String

Converts the head of the string to uppercase, and the tail to lowercase. This function is used when converting a String into an EntryType or AttrType.

Examples:

"hello" = "Hello"

"HElLo" = "Hello"