Main

Description

This module provides the 'File Input - File Output mode' functionality. It calls the parseBibTexString, ioCheckBibTex and prettyPrint functions to produce HTML that is then written into a provided file location.

Documentation

main :: IO ()

Given two file paths as arguments, read the contents of the first file, run the run function on it, and write the result into the second file.

run :: String -> IO String

Given a String, run parseBibTexString, ioCheckBibTex and prettyPrint, respectively, to produce pretty-printed HTML.