ParseBib

Description

This module provides basic parsing functionality to parse a BibTex file to a StrBibTex datatype. Does not use Components for pipe-passing mode because of its simplicity.

Documentation

main :: IO ()

Reads all the contents from standard input, attempts to parse it and convert it to an ATerm, and prints out a textual version of the ATerm on standard output.

parseBibTexString :: String -> IO StrBibTex

Try to parse a String to an StrBibTex datatype.

Datatype combinators

pBibTex :: Parser Char StrBibTex

Parse a String to a BibTex datatype. The String can be a list of entries with possible whitespace inbetween.

pEntry :: Parser Char StrEntry

Parse a single BibTex entry.

pEntryType :: Parser Char StrEntryType

Parse the type of an Entry as a String.

pEntryTitle :: Parser Char StrEntryTitle

Parse the title of an Entry as a String.

pAttr :: Parser Char StrAttr

Parses an attribute, consisting of an AttrType and AttrValue.

pAttrType :: Parser Char StrAttrType

Parses an AttrType.

pAttrValue :: Parser Char StrAttrValue

Parses an AttrValue.

Helper combinators

pQuote :: Parser Char Char

Parses a double quote.

pWhiteSpace :: Parser Char [Char]

Parses whitespace (spaces, newlines, returns and tabs).

pAlphaNum :: Parser Char Char

A parser for alphanumeric and special characters.

pAlphaNumWithBrace :: Parser Char Char

A parser extending pAlphaNum with { and }.

pEitherQuotedOrNot :: Parser Char a -> Parser Char a -> Parser Char a

Possibly quoted parser p or parser q.

pMaybeQuoted :: Parser Char a -> Parser Char a

Possibly quoted parser p or p.

pWrapWhitespace :: Parser Char a -> Parser Char a

Parser with whitespace at the beginning and/or end or none at all.

pAnyWordWithSpaces :: Parser Char [Char]

Parser of a pAlphaNum with spaces/comma's.

pAnyWordWithSpacesAndBraces :: Parser Char [Char]

Parser with spaces and braces.

pMQWordWithSpaces :: Parser Char [Char]

Either quoted parser with braces or a word with spaces.

pAnyWord :: Parser Char [Char]

Parse any pAlphaNum word.